Deploy with docker-compose
