+++ backend/templates/tasks/show.html
@@ -195,9 +195,10 @@
           </form>
         </div>
 
-        <p>
+        <!-- TODO deprecate -->
+        <!-- <p>
           <a href="{{ url_for('admin_view_ar_logs') }}" target="_blank">View Annotation Worker Logs</a>
-        </p>
+        </p> -->
 
         {% if status_assign_jobs is defined and status_assign_jobs | length > 0 %}
         <i class='small'>Refresh page to see updates</i>
@@ -623,7 +624,8 @@
               <button type="submit" class="btn btn-primary click-once">Train new model for "{{label}}"</button>
             </form>
             <p>
-              <a href="{{ url_for('admin_view_train_logs') }}" target="_blank">View Train Worker Logs</a>
+              <!-- TODO deprecate -->
+              <!-- <a href="{{ url_for('admin_view_train_logs') }}" target="_blank">View Train Worker Logs</a> -->
             </p>
 
           </div>
